Ensure same PHPUnit version for all supported PHP versions
On Travis CI, PHPUnit 6.x would be installed for PHP 7.0,
and PHPUnit 7.x for all newer PHP versions.

If using a newer PHP version on a developer machine, features not
supported in PHPUnit 6.x would not be noticed until the tests
fail on Travis CI.

We can prevent this by fixing the used PHPUnit release.