Apply leeway to all checked timestamps
The "validSince" field on a user record marks the boundary before which
Firebase ID tokens are considered revoked. This value is compared to
the "auth_time" claim of a given ID token.

Until now, we applied the leeway to the "auth_time" value, but not to
the "validSince" value, which means that, if both would contain the
value `300`, the auth_time would be modified to `0`, but validSince
would remain `300`, resulting in the token being considered revoked
although it' not.