Update `beste/clock` and `kreait/firebase-tokens`
Both are now based on the stella-maris/clock library, which provides
an interface based on the currently proposed status of PSR20. Due to
the inactivity of the PSR-20 working group this is a way to already
provide interoperability while still maintaining forward compatibility.
When the current status of PSR-20 will be released at one point in time
the stella-maris/clock package will extend the PSR-20 interface so that
this package becomes immeadiately PSR-20 compatible without any further
work necessary.