Refactor retrieval of Firebase credentials
It is not possible anymore to disable credential auto-discover.
Honestly, I don't know anymore why I added it in the first place -
discovery is automatically disabled when a Service Account is
explicitly provided.

Also, the fancy validation of the given service account credentials
have been removed as well; we pass everything to the Google Auth
Library anyway - let them handle this stuff.

Finally, custom token generation has been deferred to the Google Auth
Library as well - most of the credential types are able to create them
on their own, depending on the given credentials. That should make it
more flexible in case new environments arise: in any case, it reduces
the mental load of the Factory… which is my mental load as well.

🥳