Link to updated non-gRPC firebase package
This is a proposal to link to an updated version of ahsakharti/firestore-php.  The ahsakharti/firestore-php  (https://github.com/ahsankhatri/firestore-php) package has been archived, and it is difficult to install.  I've forked the original project and updated it to support Guzzle 7 and PHP 7/8. I fixed a few minor bugs for php 8 compatibility, and improved the documentation.  It has been published https://packagist.org and is installable with `composer require bensontrent/firestore-php`.